class_project/MSML610/Fall2025/Projects/UmdTask88_Fall2025_DGL_Fraud_Detection_in_Credit_Card_Transactions/requirements.txt
lightgbm
tqdm
matplotlib
ruff
mypy
types-PyYAML
pandas==2.2.2
pyarrow==17.0.0
numpy==1.26.4
scikit-learn==1.5.1
pyyaml==6.0.2
torch==2.0.1
dgl==2.1.0
torchdata==0.6.1
torch-geometric
pydantic<2
notebook
ipywidgets